requirements.txt
biopython==1.72
numpy==1.15.1
pandas==0.23.4
python-dateutil==2.7.4
pytz==2018.6
six==1.11.0
